arm64: dts: ti: k3-am62: Support Main UART wakeup
This series adds wakeup support for the Main UART in the device tree of
the TI AM62 family of devices. It defines the specific pins and pinctrl
states needed to wakeup the system from the Main UART via I/O
daisy-chaining. The wakeup-source property is configured to describe the
low power modes the system can wakeup from using the Main UART.

Dependencies
------------
This series is dependent on these series to be merged into the kernel:
[1], [2], and [3].

Series [1] adds the system idle states that are available on
the SoCs which are used when configuring the wakeup-source property.
Series [2] to be merged integrates the PIN_WKUP_EN macro which enables the
WKUP_EN bit for UART wakeup. Series [3] includes the DT bindings needed
for the pinctrl state wakeup implementation.

Implementation
--------------
This series is intended to be implemented along with the following
series:

1. "pmdomain: ti_sci: Handle wakeup constraint if device has pinctrl
   wakeup state" [4]: Patch which skips setting constraints for wakeup
   sources that use pinctrl state 'wakeup'.

2. "serial: 8250: omap: Add wakeup support" [5]: Implements
   wakeup from the UARTs for TI K3 SoCs

3. "arm64: dts: ti: k3-am62: Support Main UART wakeup": (this series)
   implements the functionality to wakeup the system from the Main UART

Testing
-------
Tested on a AM62P SK EVM board with all series and dependencies
implemented. Suspend/resume verified with the Main UART wakeup source
by entering a keypress on the console.

This github branch [6] has all the necessary patches to test the series
using linux-next.

Links
-----
[1] https://lore.kernel.org/linux-arm-kernel/20250812-topic-am62-dt-partialio-v6-15-v2-2-25352364a0ac@baylibre.com/
[2] https://lore.kernel.org/all/20250904112538.529857-4-a-kaur@ti.com/
[3] https://lore.kernel.org/all/20250904212455.3729029-1-k-willis@ti.com/
[4] https://github.com/kwillis01/linux/tree/b4/uart-daisy-chain-pmdomain
[5] https://github.com/kwillis01/linux/tree/b4/uart-daisy-chain-8250-omap
[6] https://github.com/kwillis01/linux/tree/uart-daisy-chain

Kendall Willis (3):
  arm64: dts: ti: k3-am62x-sk-common: Enable Main UART wakeup
  arm64: dts: ti: k3-am62a7-sk: Enable Main UART wakeup
  arm64: dts: ti: k3-am62p5-sk: Enable Main UART wakeup

 arch/arm64/boot/dts/ti/k3-am62a7-sk.dts       | 24 +++++++++++++++----
 arch/arm64/boot/dts/ti/k3-am62p5-sk.dts       | 24 +++++++++++++++----
 .../arm64/boot/dts/ti/k3-am62x-sk-common.dtsi | 24 +++++++++++++++----
 3 files changed, 60 insertions(+), 12 deletions(-)

--
2.34.1

To: Nishanth Menon <nm@ti.com>
To: Vignesh Raghavendra <vigneshr@ti.com>
To: Tero Kristo <kristo@kernel.org>
To: Rob Herring <robh@kernel.org>
To: Krzysztof Kozlowski <krzk+dt@kernel.org>
To: Conor Dooley <conor+dt@kernel.org>
Cc: linux-arm-kernel@lists.infradead.org
Cc: devicetree@vger.kernel.org
Cc: linux-kernel@vger.kernel.org
Cc: d-gole@ti.com
Cc: vishalm@ti.com
Cc: sebin.francis@ti.com
Cc: msp@baylibre.com
Cc: khilman@baylibre.com
Cc: a-kaur@ti.com
Cc: k-willis@ti.com
base-commit: 4ac65880ebca1b68495bd8704263b26c050ac010

--- b4-submit-tracking ---
{
  "series": {
    "revision": 2,
    "change-id": "20250910-uart-daisy-chain-dts-99dcc09da23e",
    "prefixes": [],
    "from-thread": "20250904212827.3730314-1-k-willis@ti.com",
    "prerequisites": [
      "message-id: <20250909044108.2541534-1-a-kaur@ti.com>",
      "message-id: <20250904212455.3729029-1-k-willis@ti.com>",
      "patch-id: 707e4619d7b844e1b67ccde28b4484b7b0d27daa",
      "patch-id: 412feb5d3fc125e489d11b726c0b2d8fd5ff36f3",
      "patch-id: 4e3ed635267f9548b9ac7a9e440971bb1e54a5ee",
      "base-commit: 5f540c4aade9f"
    ]
  }
}